+++ src/fr/gemao/view/location/AjouterChequeLocationServlet.java
@@ -61,7 +61,7 @@ public class AjouterChequeLocationServlet extends HttpServlet {
 		String numeroCheque = chequeForm.getnumeroCheque();
 		String dateEncaissement = chequeForm.getdateEncaissement();
 		
-		ChequeLocation cheque = new ChequeLocation(location, datePaiement, Float.parseFloat(montantCheque), Integer.parseInt(numeroCheque), dateEncaissement);
+		ChequeLocation cheque = new ChequeLocation(location, datePaiement, Float.parseFloat(montantCheque), Long.parseLong(numeroCheque), dateEncaissement);
 		
 		session.setAttribute("cheque", cheque);
